requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.1.0
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
pypdf>=3.17.0

